Initial commit
+ Sets up basic Gem format & adds TrixVideoFilter class to inherit from
TextFilter.

This project inherits from [html-pipeline](https://github.com/jch/html-pipeline)
and is intended to provide a filter that will take
[Trix](https://github.com/basecamp/trix)-produced content as input, find
`<figure>` tags that hold preview images for YouTube or Vimeo videos, and
convert them into a YouTube or Vimeo video URL.

The intention is that this returned URL can then be filtered through
[html-pipeline-vimeo](https://github.com/dlackty/html-pipeline-vimeo) and
[html-pipeline-youtube](https://github.com/st0012/html-pipeline-youtube) in
order to display embedded videos in Trix-produced content.

See README for explanation of why this is necessary.