Adds filter and tests it
This commit adds some extra libraries to help with debugging - `pry` and
`awesome_print`.

It props up the test suite with some basic assertions about
what the filtering should achieve and provides some helper methods for comparing
HTML snippets to each other without worrying about escaped double quotes or
other irrelevant differences.

It adds the filter itself, which works by finding all figures in a given text,
and then iterating over those figures. It looks at the source for the first
image for the given figure and determines whether it is a YouTube or Vimeo
video, and if it is, grabs the service (YouTube or Vimeo) and the id of the
video in order to construct the expected URL to that video. It then replaces the
entire figure with the URL.

It leaves everything else alone.